requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.0.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
python-dotenv>=1.0.0
pypdf>=3.0.0
docx2txt>=0.8